% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicBeta.R
\name{DynamicBeta}
\alias{DynamicBeta}
\title{Time Varying Beta via Kalman filter & smoother}
\usage{
DynamicBeta(csvfilename)
}
\arguments{
\item{csvfilename}{the name of csv file containing the track record of the fund & the benchmark}
}
\value{
A list of beta values based on Kalman Filter & smoother and the respective covariance matrices
}
\description{
Calculates the beta of an investment strategy or stock by applying the Kalman filter & smoother. Apart from the beta timeseries, the state covariances are also returned so as
to provide an estimate of the uncertainty of the results. The python package "Pykalman" is used for the calculations given its proven stability.
}
\examples{

## calling DynamicBeta() without an argument loads a test file containing an 
## example fund & index track record
dyn_beta_values = DynamicBeta() 

}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
