% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equity.R
\name{Equity-class}
\alias{Equity}
\title{Equity Class}
\arguments{
\item{Notional}{The notional amount of the trade}

\item{MTM}{The mark-to-market valuation of the trade}

\item{Currency}{The currency set that the trade belongs to}

\item{Si}{The number of years that the trade will take to start (zero if already started)}

\item{BuySell}{Takes the values of either 'Buy' or 'Sell'}

\item{ISIN}{the ISIN of the Equity}

\item{traded_price}{the price that trade was done}
}
\value{
An object of type EquityIndexFuture
}
\description{
Creates an Equity object
}
\examples{


tr1 = Equity(Notional=10000,MtM=30,Currency="EUR",Si=0,BuySell='Buy',
traded_price = 10,ISIN = "XS04340432")
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}

