% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurojackpotExample.R
\name{EuroJackpotExample}
\alias{EuroJackpotExample}
\title{Eurojackpot analysis example}
\usage{
EuroJackpotExample()
}
\value{
The final results
}
\description{
Displays how the functionality related to the eurojackpot analysis can be utilized
}
\examples{

final_results = EuroJackpotExample()
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
