% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuroJackpotResults.R
\name{EuroJackpotResults}
\alias{EuroJackpotResults}
\title{Returns all the EuroJackpot results until the end of 2023}
\usage{
EuroJackpotResults()
}
\value{
A dataframe with all the EuroJackpot results
}
\description{
Returns all the EuroJackpot results since the first draw on Feb 2004 until the end of 2023
}
\examples{

eurojackpot_results = EuroJackpotResults()
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
