% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuroLotteryAllCombinations.R
\name{EuroLotteryAllCombinations}
\alias{EuroLotteryAllCombinations}
\title{Returns all the possible number combinations for EuroMillions/EuroJackpot}
\usage{
EuroLotteryAllCombinations()
}
\value{
PnL figures
}
\description{
Returns all the possible number combinations for EuroMillions/EuroJackpot
}
\examples{

# returns all the 139,838,160 possible combinations, can create memory issues.
# all_combinations = EuroLotteryAllCombinations()
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
