% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euromillionsExample.R
\name{EuroMillionsExample}
\alias{EuroMillionsExample}
\title{Euromillions analysis example}
\usage{
EuroMillionsExample()
}
\value{
The final results
}
\description{
Displays how the functionality related to the euromillions analysis can be utilized
}
\examples{

final_results = EuroMillionsExample()
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
