% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuroMillionsResults.R
\name{EuroMillionsResults}
\alias{EuroMillionsResults}
\title{Returns all the EuroMillions results until the end of 2023}
\usage{
EuroMillionsResults()
}
\value{
A dataframe with all the EuroMillions results
}
\description{
Returns all the EuroMillions results since the first draw on Feb 2004 until the end of 2023
}
\examples{

euromillions_results = EuroMillionsResults()
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
