% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTradeDetails.R
\name{GetTradeDetails}
\alias{GetTradeDetails}
\title{Returns a list with the populated fields of a Trade Object}
\usage{
GetTradeDetails(trade)
}
\arguments{
\item{trade}{A trade Object}
}
\value{
A list of fields
}
\description{
Returns a list with the populated fields of a Trade Object
}
\examples{

example_trades = ParseTrades()
Equity_Index_Future_trade = example_trades[[18]]
populated_fields = GetTradeDetails(Equity_Index_Future_trade)
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
