% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRD.R
\docType{class}
\name{IRDFuture-class}
\alias{IRDFuture}
\title{IRD Future Class}
\arguments{
\item{Notional}{The notional amount of the trade}

\item{MTM}{The mark-to-market valuation of the trade}

\item{Currency}{The currency set that the trade belongs to}

\item{Si}{The number of years that the trade will take to start (zero if already started)}

\item{Ei}{The number of years that the trade will expire}

\item{BuySell}{Takes the values of either 'Buy' or 'Sell'}
}
\value{
An object of type IRDFuture
}
\description{
Creates an IRD Future Object with the relevant info needed to calculate the Exposure-at-Default (EAD)
}

