% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectDerivatives.R
\name{SelectDerivatives}
\alias{SelectDerivatives}
\title{Select the derivatives out of a trades' list}
\usage{
SelectDerivatives(trades_list)
}
\arguments{
\item{trades_list}{the file holding the trades of the portfolio}
}
\value{
The derivatives out of a trades' list
}
\description{
Select the derivatives out of a trades' list which will be utilized to calculate the CCR Exposure.
}
\references{
Regulation (EU) 2019/876 of the European Parliament and of the Council of 20 May 2019
http://data.europa.eu/eli/reg/2019/876/oj
}
\author{
Tasos Grivas <info@openriskcalculator.com>
}
