% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetForLifeExample.R
\name{SetForLifeExample}
\alias{SetForLifeExample}
\title{Set For Life analysis example}
\usage{
SetForLifeExample()
}
\value{
The final results
}
\description{
Displays how the functionality related to the Set For Life analysis can be utilized
}
\examples{

# This software is covered by GPL license and provided strictly for educational
# reasons (no actual investment/betting decisions should be taken based on this)

final_results = SetForLifeExample()
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
