% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetForLifeResults.R
\name{SetForLifeResults}
\alias{SetForLifeResults}
\title{Returns all the EuroJackpot results until the end of 2023}
\usage{
SetForLifeResults()
}
\value{
A dataframe with all the EuroJackpot results
}
\description{
Returns all the SetForLifeResults results since the first draw on Feb 2004 until the end of 2023
}
\examples{

eurojackpot_results = EuroJackpotResults()
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
