% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKLotteryBacktesting.R
\name{UKLotteryBacktesting}
\alias{UKLotteryBacktesting}
\title{UKLottery Backtesting}
\usage{
UKLotteryBacktesting(uklottery_results, date_since, user_input)
}
\arguments{
\item{uklottery_results}{The full list of UKLottery results}

\item{date_since}{The date after which the analysis is to be performed, i.e. 2022-12-22}

\item{user_input}{The seven numbers the user has selected}
}
\value{
The backtested results
}
\description{
Backtests the numbers the user has selected against the full (or the specified) history of UKLottery results
}
\examples{

uklottery_results = UKLotteryResults()
user_input = c(5,10,20,30,40,50)
backtested_results = UKLotteryBacktesting(uklottery_results, '2005-01-01', user_input)
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
