% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKLotteryExample.R
\name{UKLotteryExample}
\alias{UKLotteryExample}
\title{UK Lottery analysis example}
\usage{
UKLotteryExample()
}
\value{
The final results
}
\description{
Displays how the functionality related to the UK Lottery analysis can be utilized
}
\examples{

# This software is covered by GPL license and provided strictly for educational
# reasons (no actual investment/betting decisions should be taken based on this)

final_results = UKLotteryExample()
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
