% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKLotteryResults.R
\name{UKLotteryResults}
\alias{UKLotteryResults}
\title{Returns all the UKLottery results until the beginning of 2025}
\usage{
UKLotteryResults()
}
\value{
A dataframe with all the UKLottery results
}
\description{
Returns all the UKLottery results since the first draw on Nov 1994 until the beginning of 2025
}
\examples{

UKLotto_results = UKLotteryResults()
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
