% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKThunderBallBacktesting.R
\name{UKThunderballBacktesting}
\alias{UKThunderballBacktesting}
\title{UK ThunderBall Backtesting}
\usage{
UKThunderballBacktesting(ukthunderball_results, date_since, user_input)
}
\arguments{
\item{ukthunderball_results}{The full list of UK ThunderBall results}

\item{date_since}{The date after which the analysis is to be performed, i.e. 2022-12-22}

\item{user_input}{The seven numbers the user has selected}
}
\value{
The backtested results
}
\description{
Backtests the numbers the user has selected against the full (or the specified) history of UK ThunderBall results
}
\examples{

ukthunderball_results = UKThunderBallResults()
user_input = c(10,20,30,31,32,5)
backtested_results = UKThunderballBacktesting(ukthunderball_results, user_input = user_input)
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
