% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKThunderBallExample.R
\name{UKThunderballExample}
\alias{UKThunderballExample}
\title{UK ThunderBall analysis example}
\usage{
UKThunderballExample()
}
\value{
The final results
}
\description{
Displays how the functionality related to the UK ThunderBall analysis can be utilized
}
\examples{

# This software is covered by GPL license and provided strictly for educational
# reasons (no actual investment/betting decisions should be taken based on this)

final_results = EuroJackpotExample()
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
