% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKThunderBallResults.R
\name{UKThunderBallResults}
\alias{UKThunderBallResults}
\title{Returns all the EuroJackpot results until the end of 2023}
\usage{
UKThunderBallResults()
}
\value{
A dataframe with all the EuroJackpot results
}
\description{
Returns all the EuroJackpot results since the first draw on Feb 2004 until the end of 2023
}
\examples{

UKThunderBall_Results = UKThunderBallResults()
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
