% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKThunderBallResults.R
\name{UKThunderBallResults}
\alias{UKThunderBallResults}
\title{Returns all the UK ThunderBall results until the beginning of 2025}
\usage{
UKThunderBallResults()
}
\value{
A dataframe with all the UK ThunderBall results
}
\description{
Returns all the UK ThunderBall results until the beginning of 2025
}
\examples{

UKThunderBall_Results = UKThunderBallResults()
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
